Check for null after calling listFiles() in InstanceProvider#deleteAllFilesInDirectory()

+++ collect_app/src/main/java/org/odk/collect/android/provider/InstanceProvider.java
@@ -224,10 +224,12 @@ public class InstanceProvider extends ContentProvider {
 
                 // delete all the files in the directory
                 File[] files = directory.listFiles();
-                for (File f : files) {
-                    // should make this recursive if we get worried about
-                    // the media directory containing directories
-                    f.delete();
+                if (files != null) {
+                    for (File f : files) {
+                        // should make this recursive if we get worried about
+                        // the media directory containing directories
+                        f.delete();
+                    }
                 }
             }
             directory.delete();
